mmngr_drv: mmngr: handle_invalid_translation
This fix is based on https://github.com/renesas-rcar/mmngr_drv/pull/5
When IPMMU support is enabled and address transformation involves a translation table,
the input physical address may not match any section of the transformation table.
In such case, the default zero value of the virtual address is returned and propagated to the user.
The consequences of this error might be visible only in the userspace and finding the root cause may require some effort.
It is proposed to handle translation faults as earlier as possible and return an error to a client.

Signed-off-by: NguyenB Nguyen <nguyen.nguyen.yj@renesas.com>